moving FAQ to home page
